requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==2.2.8
django-neomodel==0.1.0
djangorestframework==3.13.1
idna==3.4
interchange==2021.0.4
monotonic==1.6
neo4j-driver==4.4.10
neobolt==1.7.17
neomodel==5.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pansi==2020.7.3
py2neo==2021.2.3
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
